read manifest file size as 64 bits
The file-size field is eight bytes on disk; it was read as four, which returns the
correct value for every file below 4 GiB and silently wraps above it. Nothing had
failed yet because League's largest shipped file is 3.11 GiB - inside the range
where both widths agree.

The FlatBuffer vtable settles the width. In a real file entry the eight-byte
fields are packed together: id at 16, directory id at 24, size at 32, locale
flags at 40. Size therefore owns eight bytes; were it four, the flags would sit
at 36.

The four bytes never read are the high half, so the wrap would have been silent -
a wrong number with no error, trusted by anything doing download or patch
arithmetic.

The existing real-file test already cross-checks this without knowing it: a
file's uncompressed chunk sizes must sum to its declared size, and that holds
across all three manifests with the wider read. Added a test pinning the largest
known entry at 3337364776 bytes exactly, and a u64 binding that fails to compile
if the field is ever narrowed again.

Two synthetic fixtures hand-encoded a four-byte size, baking in the same wrong
assumption; their field offsets now match the real layout.

+++ crates/rs_rman/src/rman.rs
@@ -56,7 +56,9 @@ pub struct Bundle {
 pub struct FileEntry {
     pub id: u64,
     pub name: String,
-    pub size: u32,
+    /// Uncompressed size in bytes. 64-bit on disk: League already ships files above 3 GiB, so a
+    /// 32-bit read would silently wrap once one crosses 4 GiB.
+    pub size: u64,
     pub directory_id: Option<u64>,
     pub chunk_ids: Vec<u64>,
     pub link: Option<String>,
@@ -171,7 +173,7 @@ impl Rman {
                     }
                     None => f.name.clone(),
                 };
-                (path, f.size as u64)
+                (path, f.size)
             })
             .collect()
     }
